feat: add eslint, prettier & ts
